requirements.txt
pandas==3.0.1
playwright==1.58.0
requests==2.32.5
openpyxl==3.1.5
lxml==6.0.2
python-dotenv==1.2.2